+++ comb_spec_searcher/strategies/rule.py
@@ -322,7 +322,7 @@ class AbstractRule(abc.ABC, Generic[CombinatorialClassType, CombinatorialObjectT
             symbol_height = min(1, len(res) - 1)
             eq_symbol = (
                 ["     " for i in range(symbol_height)]
-                + [f"  {self.get_op_symbol()}  "]
+                + [f"  {self.get_eq_symbol()}  "]
                 + ["     " for i in range(symbol_height)]
             )
             join(res, eq_symbol)
@@ -818,6 +818,8 @@ class EquivalencePathRule(Rule[CombinatorialClassType, CombinatorialObjectType])
                 original_constructor = rule.constructor
                 assert isinstance(original_constructor, (DisjointUnion, Complement))
                 rules_parameters = original_constructor.extra_parameters[0]
+                if isinstance(original_constructor, Complement):
+                    rules_parameters = {b: a for a, b in rules_parameters.items()}
                 extra_parameters = {
                     parent_var: rules_parameters[child_var]
                     for parent_var, child_var in extra_parameters.items()