+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   not have
 - Trying to verify the class being expanded when expanding a combinatorial
   specification.
+- Bug in complement constructors inside equivalence path rules.
 
 ## [4.0.0] - 2021-06-14
 ### Added